Split classes in 3 files
